Replace git-whatchanged usage with git-log
git-whatchanged has been deprecated in Git 2.51 and now refuses to run
without an override flag specified. According to the git-whatchanged
documentation, it "is essentially the same as git-log(1) but defaults to
showing the raw format diff output and skipping merges".

This patch replaces the usage of the whatchanged sub-command with log,
passing --raw to preserve compatibility with previous behavior.